prepare submission
