--- bachelor_controller/config_master.yaml
@@ -1,28 +1,0 @@
-## UniFlex Agent config file
-
-config:
-  name: "Global_Controller"
-  info: 'agent hosts global controller'
-  iface: 'lo'
-  sub: "tcp://127.0.0.1:8990"
-  pub: "tcp://127.0.0.1:8989"
-
-broker:
-  xpub: "tcp://127.0.0.1:8990"
-  xsub: "tcp://127.0.0.1:8989"
-
-control_applications:
-  myController:
-      file : my_control_app.py
-      class_name : MyController
-      kwargs : {}
-
-modules:
-  discovery:
-        module : uniflex_app_discovery_pyre
-        class_name : PyreDiscoveryMasterModule
-        kwargs: {"iface":"lo",
-                 "groupName":"uniflex_1234",
-                 "sub":"tcp://127.0.0.1:8990",
-                 "pub":"tcp://127.0.0.1:8989"
-                }
