--- bachelor_controller/config_slave2.yaml
@@ -1,25 +1,0 @@
-## UniFlex Agent config file
-
-config:
-  name: 'HC node'
-  info: 'filter runs on local node'
-  iface: 'lo'
-
-control_applications:
-  myFilter:
-      file : my_filter.py
-      class_name : MyAvgFilter
-      kwargs : {"window": 5}
-
-  discovery:
-      module : uniflex_app_discovery_pyre
-      class_name : PyreDiscoverySlaveModule
-      kwargs: {"iface":"lo", "groupName":"uniflex_1234"}
-
-modules:
-  simple:
-      module : uniflex_module_simple
-      class_name : SimpleModule4
-      devices : ['phy0']
-      kwargs : { "MAC_List" : ["38:10:d5:9a:0b:60"]}
-
