+++ openAI_RRM/SimulationSlavesConfig/Setting2_2unsym/config_slave2.yaml
@@ -2,7 +2,7 @@
 
 config:
   name: 'HC node'
-  info: 'filter runs on local node'
+  info: 'rrm control simulation'
   iface: 'lo'
   iface: 'lo'
   sub: "tcp://127.0.0.1:8990"
@@ -13,11 +13,6 @@ config:
 #  xpub: "tcp://127.0.0.1:8989"
 
 control_applications:
-  myFilter:
-      file : my_filter.py
-      class_name : MyAvgFilter
-      kwargs : {"window": 5}
-
   discovery:
       module : uniflex_app_discovery_pyre
       class_name : PyreDiscoverySlaveModule
