--- openAI_RRM/SimulationSlavesConfig/Setting3_varSetsmall/config_slave.yaml
@@ -1,38 +1,0 @@
-## UniFlex Agent config file
-
-config:
-  name: 'HC node'
-  info: 'filter runs on local node'
-  iface: 'lo'
-  iface: 'lo'
-  sub: "tcp://127.0.0.1:8990"
-  pub: "tcp://127.0.0.1:8989"
-
-broker:
-  xsub: "tcp://127.0.0.1:8990"
-  xpub: "tcp://127.0.0.1:8989"
-
-control_applications:
-  myFilter:
-      file : my_filter.py
-      class_name : MyAvgFilter
-      kwargs : {"window": 5}
-
-  discovery:
-      module : uniflex_app_discovery_pyre
-      class_name : PyreDiscoverySlaveModule
-      kwargs: {"iface":"lo", "groupName":"uniflex_1234"}
-
-modules:
-  simple:
-      module : uniflex_module_simple
-      class_name : SimpleModule4
-      devices : ['phy0']
-      kwargs : { "clients" : ["cc:cc:cc:cc:cc:01", "cc:cc:cc:cc:cc:02", "cc:cc:cc:cc:cc:03", "cc:cc:cc:cc:cc:04", "cc:cc:cc:cc:cc:05", "cc:cc:cc:cc:cc:06"], 
-        'neighbors' : [['aa:aa:aa:aa:aa:02', 'aa:aa:aa:aa:aa:03'], ['aa:aa:aa:aa:aa:02', 'aa:aa:aa:aa:aa:03'], ['aa:aa:aa:aa:aa:02'], ['aa:aa:aa:aa:aa:02'], ['aa:aa:aa:aa:aa:03']],
-        'myMAC' : 'aa:aa:aa:aa:aa:01',
-        'simulation': {'channelSwitchingTime': 0, 'channelThroughputDefault':54000000, 'txBytesRandom':0, 'clientnum':1, 'clientconf': '/home/sascha/tu-cloud/Uni/Module/Bachelorarbeit_TI/Messungsautomatisierung/clientconfig/3_112_AP1',
-            'mode': 'training', # training or working
-            'numsClients': [1,1,1,1,1]
-        }}
-
