--- openAI_RRM/SimulationSlavesConfig/Setting3_varSetsmall/config_slave3.yaml
@@ -1,38 +1,0 @@
-## UniFlex Agent config file
-
-config:
-  name: 'HC node'
-  info: 'filter runs on local node'
-  iface: 'lo'
-  iface: 'lo'
-  sub: "tcp://127.0.0.1:8990"
-  pub: "tcp://127.0.0.1:8989"
-
-broker:
-  xsub: "tcp://127.0.0.1:8990"
-  xpub: "tcp://127.0.0.1:8989"
-
-control_applications:
-  myFilter:
-      file : my_filter.py
-      class_name : MyAvgFilter
-      kwargs : {"window": 5}
-
-  discovery:
-      module : uniflex_app_discovery_pyre
-      class_name : PyreDiscoverySlaveModule
-      kwargs: {"iface":"lo", "groupName":"uniflex_1234"}
-
-modules:
-  simple:
-      module : uniflex_module_simple
-      class_name : SimpleModule4
-      devices : ['phy0']
-      kwargs : { "clients" : ["cc:cc:cc:cc:cc:21", "cc:cc:cc:cc:cc:22", "cc:cc:cc:cc:cc:23", "cc:cc:cc:cc:cc:24", "cc:cc:cc:cc:cc:25", "cc:cc:cc:cc:cc:26"],
-        'neighbors' : [['aa:aa:aa:aa:aa:02', 'aa:aa:aa:aa:aa:01'], ['aa:aa:aa:aa:aa:02', 'aa:aa:aa:aa:aa:01'], ['aa:aa:aa:aa:aa:02'], ['aa:aa:aa:aa:aa:02'], ['aa:aa:aa:aa:aa:01']],
-        'myMAC' : 'aa:aa:aa:aa:aa:03',
-        'simulation': {'channelSwitchingTime': 0, 'channelThroughputDefault':54000000, 'txBytesRandom':0, 'clientnum':2, 'clientconf': '/home/sascha/tu-cloud/Uni/Module/Bachelorarbeit_TI/Messungsautomatisierung/clientconfig/3_112_AP3', 
-        'mode': 'training', # training or working
-        'numsClients': [2,2,1,1,1]
-        }}
-
