+++ openAI_RRM/config_slave.yaml
@@ -8,9 +8,9 @@ config:
   sub: "tcp://192.168.10.157:8990"
   pub: "tcp://192.168.10.157:8989"
 
-broker:
-  xpub: "tcp://192.168.10.157:8990"
-  xsub: "tcp://192.168.10.157:8989"
+#broker:
+#  xpub: "tcp://192.168.10.157:8990"
+#  xsub: "tcp://192.168.10.157:8989"
 
 control_applications:
   myFilter:
