+++ openAI_RRM/readme.txt
@@ -6,12 +6,13 @@ uniflex-broker
 # 2b. Run control program in master node:
 python3 rrm_agent.py --config ./config_master.yaml
 python3 rrm_agent.py --config ./config_master_simulation.yaml
+
 # 2c. Run modules in slave node:
 #Linux WiFi AP
 uniflex-agent --config ./config_slave.yaml
 #Simulation
-uniflex-agent --config ./SimulationSlavesConfig/Setting2_2/config_slave.yaml
-uniflex-agent --config ./SimulationSlavesConfig/Setting2_2/config_slave2.yaml
-uniflex-agent --config ./SimulationSlavesConfig/Setting2_2/config_slave3.yaml
+uniflex-agent --config ./SimulationSlavesConfig/##Name of Experiment##/config_slave.yaml
+uniflex-agent --config ./SimulationSlavesConfig/##Name of Experiment##/config_slave2.yaml
+# and so on
 
 # For debugging mode run with -v option
