+++ openAI_RRM/rl_agent_multi_evalnetwork.py
@@ -1,6 +1,8 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 
+##TODO: Insert path to neuronal network
+
 import gym
 import UniFlexGym
 import tensorflow as tf
@@ -20,7 +22,81 @@ sortedIndecies = []
 ac_space = []
 BANDWITH_ON_CHANNEL = 54e6
 numChannels = 2
-SORT_VALUES = False
+SORT_VALUES = True
+topologies = [[1,2,1], [2,2,2], [1,0,1]]
+N_test = int(1000)
+aps = 3
+maxclients = 100
+
+def zeros(anz):
+    return np.zeros(anz)
+
+def ones(anz):
+    return np.ones(anz)
+
+def calculate_all_best_action(clients, aps):
+    aps_sort = sorted(aps)
+    #if string topology
+    if aps_sort == [1,1,2]:
+        #set ap in the middle to 1, all other to 0
+        result = [zeros(3), ones(3)]
+        result[0][aps.index(2)] = 1
+        result[1][aps.index(2)] = 0
+        return result
+    
+    # if island topology
+    if aps_sort == [0,1,1]:
+        #set on of the neighouring aps to 1, all other to 0
+        result = []
+        group = np.where(np.array(aps) == 1)
+        for elem in group[0]:
+            myresult = zeros(3)
+            myresult[elem] = 1
+            result.append(myresult)
+            myresult = ones(3)
+            myresult[elem] = 0
+            result.append(myresult)
+        return result
+    
+    #if all aps can hear all other
+    if aps_sort == [2,2,2]:
+        result = []
+        #get ap with most clients
+        clients_sort = sorted(clients)
+        ap_most = np.where(clients == clients_sort[2])
+        for elem in ap_most[0]:
+            myresult = zeros(3)
+            myresult[elem] = 1
+            result.append(myresult)
+            myresult = ones(3)
+            myresult[elem] = 0
+            result.append(myresult)
+        return result
+    # there is no topology
+    print("notopo")
+    return [zeros(3)]
+
+#test data consist on a vector of observations and a label vector of all valid 
+# action to this observations
+def generate_testdata(number, dimension, maxclients, topoplogies, sortValues):
+    data = []
+    labels = []
+    for i in range(number):
+        for topology in topologies:
+            index = np.array(range(dimension), dtype=np.int16)
+            clients = np.random.randint(maxclients, size=dimension)
+            mydata = np.vstack((clients, topology, index)).transpose()
+            labeldata = mydata
+            #if sortValues:
+            #    labeldata = np.sort(labeldata.view('i4,i4'), order=['f0', 'f1'], axis=0).view(np.int)
+            clients = labeldata[:,0]
+            topology = labeldata[:,1].tolist()
+            #mydata = np.reshape(mydata, [1, 2*dimension])
+            mydata = np.delete(mydata, 2, axis=1)
+            data.append(mydata)
+            #labels.append(channelvectors_to_label(calculate_all_best_action(clients, topology)))
+            labels.append(calculate_all_best_action(clients, topology))
+    return [data, labels]
 
 def normalize_state(state, ob_space, s_size):
     global sortedIndecies
@@ -58,6 +134,7 @@ def eval(clients):
     errorcounter_cli = 0
     errorcounter_ap  = 0
     counter = 0
+    errorlog = ""
 
     for client in clients:
         ap = client['aps']
@@ -97,7 +174,8 @@ def eval(clients):
         #        success_ap = True
         #        break
         
-        print("[Cli, Ap]: Cli:" + str(client['clients']) + ", AP:" + str(ap) + ", Action:" +str(action) + ", Actionvector" + str(actionvector) + ", " + str(success_cli))
+        output = "[Cli, Ap]: Cli:" + str(client['clients']) + ", AP:" + str(ap) + ", Action:" +str(action) + ", Actionvector" + str(actionvector) + ", " + str(success_cli) + " SortedID: " +str(sortedIndecies)
+        print(output)
         #print("[Ap, Cli]: Cli:" + str(client['clients']) + ", AP:" + str(ap) + ", Action:" +str(actionap) + ", Actionvector" + str(actionvectorap) + ", " + str(success_ap))
         counter += 1
         
@@ -106,9 +184,11 @@ def eval(clients):
         
         if not success_cli:
             errorcounter_cli +=1
+            errorlog += output +"\n"
 
-    print("Errors in [Cli,Ap]:" + str(errorcounter_cli) + "/" + str(counter) + "(" + str(errorcounter_cli/counter) + "%)")
+    print("Errors in [Cli,Ap]:" + str(errorcounter_cli) + "/" + str(counter) + "(" + str(errorcounter_cli/counter*100) + "%)")
     #print("Errors in [Ap,Cli]:" + str(errorcounter_ap) + "/" + str(counter) + "(" + str(errorcounter_ap/counter) + "%)")
+    print(errorlog)
 
 def calculate_reward(clients_p_ap, action):
     reward = 0
@@ -189,57 +269,24 @@ def eval_handover(client, new_clients):
         
         print("Cli:" + str(new_client['clients']) + ", AP:" + str(ap) + ", Action:" +str(action) + ", Actionvector" + str(actionvector) + ", " + str(success_cli) + ", reward:" + str(reward))
 
+
 ac_space = spaces.MultiDiscrete([2,2,2])
 ob_space = spaces.Box(low=0, high=6, shape=(ac_space.nvec.shape[0],2), dtype=np.uint32)
-#trainingfileap = "/home/sascha/tu-cloud/Uni/Module/Bachelorarbeit_TI/Messungsautomatisierung/simulationMeasurements_2/test/logs/controller_3_112neuronalesNetz.train"
-trainingfile = "/home/sascha/tu-cloud/Uni/Module/Bachelorarbeit_TI/Messungsautomatisierung/simulationMeasurements_2/training4_unsort_3set_2/logs/controller_3_varSetneuronalesNetz.train"
+trainingfile = "path-to-neuronal-network.train"
 
-clients = [ {'clients': [1, 1, 5], 'aps': [2,2,2], 'valid':[[1,1,0], [0,0,1]]},
-            {'clients': [1, 3, 2], 'aps': [2,2,2], 'valid':[[0,1,0], [1,0,1]]},
-            {'clients': [5, 3, 4], 'aps': [2,2,2], 'valid':[[1,0,0], [0,1,1]]},
-            {'clients': [5, 1, 3], 'aps': [1,2,1], 'valid':[[0,1,0], [1,0,1]]},
-            {'clients': [2, 4, 2], 'aps': [1,2,1], 'valid':[[0,1,0], [1,0,1]]},
-            {'clients': [7, 1, 5], 'aps': [1,2,1], 'valid':[[0,1,0], [1,0,1]]},
-            {'clients': [4, 3, 2], 'aps': [1,0,1], 'valid':[[0,1,1], [0,0,1], [1,1,0], [1,0,0]]},
-            {'clients': [1, 3, 2], 'aps': [1,0,1], 'valid':[[0,1,1], [0,0,1], [1,1,0], [1,0,0]]}
-            ]
-#clients2 = [{'clients': [1, 1, 2], 'valid':[[1,0,1], [0,1,0]]},
-#            {'clients': [1, 2, 3], 'valid':[[1,0,1], [0,1,0]]},
-#            {'clients': [6, 0, 0], 'valid':[[1,0,1], [0,1,0]]},
-#            {'clients': [1, 5, 1], 'valid':[[1,0,1], [0,1,0]]},
-#            {'clients': [5, 1, 1], 'valid':[[1,0,1], [0,1,0]]},
-#            {'clients': [2, 2, 2], 'valid':[[1,0,1], [0,1,0]]},
-#            {'clients': [5, 5, 1], 'valid':[[1,0,1], [0,1,0]]}
-#            ]
-handover = [{'clients': [1, 5, 1], 'aps': [2,2,2], 'valid':[[1,0,1], [0,1,0]]},
-            {'clients': [2, 4, 1], 'aps': [2,2,2], 'valid':[[1,0,1], [0,1,0]]},
-            {'clients': [1, 4, 2], 'aps': [2,2,2], 'valid':[[1,0,1], [0,1,0]]}
+#generate random test data
+[test_data, test_labels] = generate_testdata(N_test, aps, maxclients, topologies, SORT_VALUES)
+
+clients = []
+for elem, label in zip(test_data, test_labels):
+    clients.append({'clients': elem[:,0], 'aps': elem[:,1], 'valid':label})
+
+
+special = [{'clients': [10, 10, 1000], 'aps': [2,2,2], 'valid':[[1,1,0], [0,0,1]]},
+            {'clients': [1000, 10, 10], 'aps': [1,2,1], 'valid':[[1,0,1], [0,1,0]]},
+            {'clients': [1000, 4, 2], 'aps': [1,0,1], 'valid':[[1,0,0], [0,0,1],[1,1,0], [0,1,1]]}
             ]
 
-handover2 = [{'clients': [1, 5, 1], 'aps': [1,2,1], 'valid':[[1,0,1], [0,1,0]]},
-             {'clients': [2, 4, 1], 'aps': [1,2,1], 'valid':[[1,0,1], [0,1,0]]},
-             {'clients': [1, 4, 2], 'aps': [1,2,1], 'valid':[[1,0,1], [0,1,0]]}
-            ]
-
-#handover2 = [{'clients': [1, 2, 1], 'valid':[[1,0,1], [0,1,0]]},
-#            {'clients': [2, 1, 1], 'valid':[[0,1,1], [1,0,0]]},
-#            {'clients': [1, 1, 2], 'valid':[[1,1,0], [0,0,1]]}
-#            ]
-
-handover3 = [{'clients': [2, 2, 1], 'aps': [1,2,1], 'valid':[[1,0,1], [0,1,0], [1,0,0], [0,1,1]]},
-             {'clients': [2, 1, 2], 'aps': [1,2,1],  'valid':[[0,1,1], [1,0,0], [1,1,0], [0,0,1]]},
-             {'clients': [1, 2, 2], 'aps': [1,2,1],  'valid':[[1,1,0], [0,0,1], [1,0,1], [0,1,0]]}
-            ]
-
-#aps  = [[2,2,2]]
-#aps2 = [[1,2,1]]
-
-#states = [[[1,2],[1,2],[2,2]], [[1,2],[2,2],[3,2]], [[6,2],[0,2],[0,2]], [[1,2],[5,2],[1,2]], [[2,2],[2,2],[2,2]],
-#          [[1,1],[1,0],[2,1]], [[1,1],[2,1],[3,0]], [[6,1],[0,0],[0,1]], [[1,1],[5,1],[1,0]], [[2,0],[2,1],[2,1]],
-#          [[2,1],[2,2],[2,1]], [[3,1],[2,2],[2,1]]]
-#states = [[[2,1],[2,1],[2,2]], [[2,1],[2,2],[2,3]], [[2,6],[2,0],[2,0]], [[2,1],[2,5],[2,1]], [[2,2],[2,2],[2,2]],
-#          [[0,1],[1,1],[1,2]], [[1,1],[0,2],[1,3]], [[1,6],[1,0],[0,0]], [[1,1],[1,5],[0,1]], [[1,2],[0,2],[1,2]],
-#          [[1,2],[2,2],[1,2]], [[1,3],[2,2],[1,2]]]
 
 print("Observation space: ", ob_space,  ob_space.dtype)
 print("Action space: ", ac_space, ac_space.nvec)
@@ -250,38 +297,20 @@ s_size = tmps_size[0] * tmps_size[1]
 a_size = pow(ac_space.nvec[0], ac_space.nvec.shape[0])
 
 print("observation_space size:" + str(s_size))
-
-print("Data: Trained Data of different settings with unsorted agent. Experiment 4")
+print("Data: Trained Data of different settings with sorting agent. Experiment 3")
 
 model = keras.Sequential()
-model.add(keras.layers.Dense(s_size, input_shape=(s_size,), activation='sigmoid'))
-model.add(keras.layers.Dense(5, activation='relu'))
+model.add(keras.layers.Dense(s_size, input_shape=(s_size,), activation='relu'))
+#model.add(keras.layers.Dense(5, activation='relu'))
 model.add(keras.layers.Dense(a_size, activation='softmax'))
 model.compile(optimizer=tf.train.AdamOptimizer(0.001),
               loss='categorical_crossentropy',
               metrics=['accuracy'])
 model.load_weights(trainingfile)
 
-#modelap = keras.Sequential()
-#modelap.add(keras.layers.Dense(s_size, input_shape=(s_size,), activation='sigmoid'))
-#modelap.add(keras.layers.Dense(5, activation='relu'))
-#modelap.add(keras.layers.Dense(a_size, activation='softmax'))
-#modelap.compile(optimizer=tf.train.AdamOptimizer(0.001),
-#              loss='categorical_crossentropy',
-#              metrics=['accuracy'])
-#modelap.load_weights(trainingfileap)
-
-print("\nSame domain:")
+print("\n3000 zufällige Szenarien:")
 eval(clients)
 
-#print("\nMan in the middle:")
-#eval(clients2, aps2)
+print("\Spezielle Szenarien:")
+eval(special)
 
-print("\nHandover test")
-eval_handover(handover[0], handover[1:])
-
-print("\nHandover test 2")
-eval_handover(handover2[0], handover2[1:])
-
-print("\nHandover test 3")
-eval_handover(handover3[0], handover3[1:])
