+++ openAI_RRM/thompson_agent2.py
